+++ packages/manager/cypress/e2e/core/loadBalancers/load-balancer-landing-page.spec.ts
@@ -1,5 +1,5 @@
 /**
- * @file Integration tests for Akamai Global Load Balancer navigation.
+ * @file Integration tests for Akamai Cloud Load Balancer navigation.
  */
 
 import {
@@ -18,24 +18,23 @@ import {
 } from 'support/intercepts/load-balancers';
 import type { Loadbalancer } from '@linode/api-v4';
 import { chooseRegion } from 'support/util/regions';
-import { getRegionById } from 'support/util/regions';
 
 /**
- * Navigates to the AGLB landing page using breadcrumb navigation.
+ * Navigates to the ACLB landing page using breadcrumb navigation.
  *
  * Asserts that the URL has updated to reflect navigation.
  */
 const returnToLandingPage = () => {
   ui.entityHeader.find().within(() => {
-    cy.findByText('Global Load Balancers').should('be.visible').click();
+    cy.findByText('Cloud Load Balancers').should('be.visible').click();
   });
 
   cy.url().should('endWith', '/loadbalancers');
 };
 
-describe('Akamai Global Load Balancer landing page', () => {
+describe('Akamai Cloud Load Balancer landing page', () => {
   /*
-   * - Confirms that load balancers are listed on the AGLB landing page.
+   * - Confirms that load balancers are listed on the ACLB landing page.
    * - Confirms that clicking a load balancer label directs to its details pages.
    * - Confirms that Create Loadbalancer button is present and enabled.
    * - Confirms that load balancer action menu items are present.
@@ -57,9 +56,9 @@ describe('Akamai Global Load Balancer landing page', () => {
       }),
     ];
 
-    // TODO Delete feature flag mocks when AGLB feature flag goes away.
+    // TODO Delete feature flag mocks when ACLB feature flag goes away.
     mockAppendFeatureFlags({
-      aglb: makeFeatureFlagData(true),
+      aclb: makeFeatureFlagData(true),
     }).as('getFeatureFlags');
     mockGetFeatureFlagClientstream().as('getClientStream');
     mockGetLoadBalancers(loadbalancerMocks).as('getLoadBalancers');
@@ -74,7 +73,7 @@ describe('Akamai Global Load Balancer landing page', () => {
         .should('be.visible')
         .closest('tr')
         .within(() => {
-          // TODO: AGLB - Confirm that regions from the API are listed for load balancer
+          // TODO: ACLB - Confirm that regions from the API are listed for load balancer
           // loadbalancerMock.regions.forEach((loadbalancerRegion: string) => {
           //   const regionLabel = getRegionById(loadbalancerRegion).label;
           //   cy.findByText(regionLabel, { exact: false }).should('be.visible');
@@ -96,7 +95,7 @@ describe('Akamai Global Load Balancer landing page', () => {
         .should('be.visible')
         .closest('tr')
         .within(() => {
-          // TODO: AGLB - Confirm that regions from the API are listed for load balancer
+          // TODO: ACLB - Confirm that regions from the API are listed for load balancer
           // loadbalancerMock.regions.forEach((loadbalancerRegion: string) => {
           //   const regionLabel = getRegionById(loadbalancerRegion).label;
           //   cy.findByText(regionLabel, { exact: false }).should('be.visible');
@@ -132,8 +131,8 @@ describe('Akamai Global Load Balancer landing page', () => {
 
 describe('Delete', () => {
   /*
-   * - Confirms that Deleting a load balancer from the AGLB landing page.
-   * - Confirms AGLB landing page reverts to its empty state when all of the load balancers have been deleted.
+   * - Confirms that Deleting a load balancer from the ACLB landing page.
+   * - Confirms ACLB landing page reverts to its empty state when all of the load balancers have been deleted.
    */
   it('Delete a Load Balancer from landing page.', () => {
     const chosenRegion = chooseRegion();
@@ -152,9 +151,9 @@ describe('Delete', () => {
       }),
     ];
 
-    // TODO Delete feature flag mocks when AGLB feature flag goes away.
+    // TODO Delete feature flag mocks when ACLB feature flag goes away.
     mockAppendFeatureFlags({
-      aglb: makeFeatureFlagData(true),
+      aclb: makeFeatureFlagData(true),
     }).as('getFeatureFlags');
     mockGetFeatureFlagClientstream().as('getClientStream');
     mockGetLoadBalancers(loadbalancerMocks).as('getLoadBalancers');
@@ -200,7 +199,7 @@ describe('Delete', () => {
 
     cy.get('[data-qa-header]')
       .should('be.visible')
-      .should('have.text', 'Global Load Balancers');
+      .should('have.text', 'Cloud Load Balancers');
 
     cy.findByText(
       'Scalable Layer 4 and Layer 7 load balancer to route and manage enterprise traffic between clients and your distributed applications and networks globally.'
@@ -209,7 +208,7 @@ describe('Delete', () => {
 
     // Create button exists and navigates user to create page.
     ui.button
-      .findByTitle('Create Global Load Balancer')
+      .findByTitle('Create Cloud Load Balancer')
       .should('be.visible')
       .should('be.enabled');
 
@@ -233,9 +232,9 @@ describe('Delete', () => {
       }),
     ];
 
-    // TODO Delete feature flag mocks when AGLB feature flag goes away.
+    // TODO Delete feature flag mocks when ACLB feature flag goes away.
     mockAppendFeatureFlags({
-      aglb: makeFeatureFlagData(true),
+      aclb: makeFeatureFlagData(true),
     }).as('getFeatureFlags');
     mockGetFeatureFlagClientstream().as('getClientStream');
     mockGetLoadBalancers(loadbalancerMocks).as('getLoadBalancers');