+++ packages/manager/cypress/support/constants/dc-specific-pricing.ts
@@ -51,6 +51,11 @@ export const dcPricingMockLinodeTypes = linodeTypeFactory.buildList(3, {
           id: 'us-west',
           monthly: 4.17,
         },
+        {
+          hourly: 0.006,
+          id: 'us-southeast',
+          monthly: 4.67,
+        },
       ],
     },
   },
@@ -103,6 +108,11 @@ export const dcPricingMockLinodeTypesForBackups = linodeTypeFactory.buildList(
             id: 'us-west',
             monthly: 4.17,
           },
+          {
+            hourly: 0.006,
+            id: 'us-southeast',
+            monthly: 4.67,
+          },
         ],
       },
     },
@@ -117,7 +127,7 @@ export const dcPricingMockLinodeTypesForBackups = linodeTypeFactory.buildList(
 export const dcPricingLkeClusterPlans: LkePlanDescription[] = dcPricingMockLinodeTypes.map(
   (type) => {
     return {
-      size: type.id.split('-')[2],
+      size: parseInt(type.id.split('-')[2], 10),
       tab: 'Shared CPU',
       type: 'Linode',
     };