+++ packages/manager/src/MainContent.tsx
@@ -30,6 +30,7 @@ import { isFeatureEnabled } from 'src/utilities/accountCapabilities';
 import { ENABLE_MAINTENANCE_MODE } from './constants';
 import { complianceUpdateContext } from './context/complianceUpdateContext';
 import { FlagSet } from './featureFlags';
+import { useIsACLBEnabled } from './features/LoadBalancers/utils';
 import { useGlobalErrors } from './hooks/useGlobalErrors';
 
 const useStyles = makeStyles()((theme: Theme) => ({
@@ -159,8 +160,10 @@ const Help = React.lazy(() =>
 const SearchLanding = React.lazy(
   () => import('src/features/Search/SearchLanding')
 );
-const EventsLanding = React.lazy(
-  () => import('src/features/Events/EventsLanding')
+const EventsLanding = React.lazy(() =>
+  import('src/features/Events/EventsLanding').then((module) => ({
+    default: module.EventsLanding,
+  }))
 );
 const AccountActivationLanding = React.lazy(
   () => import('src/components/AccountActivation/AccountActivationLanding')
@@ -169,6 +172,11 @@ const Firewalls = React.lazy(() => import('src/features/Firewalls'));
 const Databases = React.lazy(() => import('src/features/Databases'));
 const BetaRoutes = React.lazy(() => import('src/features/Betas'));
 const VPC = React.lazy(() => import('src/features/VPCs'));
+const PlacementGroups = React.lazy(() =>
+  import('src/features/PlacementGroups').then((module) => ({
+    default: module.PlacementGroups,
+  }))
+);
 
 export const MainContent = () => {
   const { classes, cx } = useStyles();
@@ -215,6 +223,8 @@ export const MainContent = () => {
     Boolean(flags.vpc),
     account?.capabilities ?? []
   );
+
+  const { isACLBEnabled } = useIsACLBEnabled();
 
   const defaultRoot = _isManagedAccount ? '/managed' : '/linodes';
 
@@ -325,9 +335,13 @@ export const MainContent = () => {
                     <React.Suspense fallback={<SuspenseLoader />}>
                       <Switch>
                         <Route component={LinodesRoutes} path="/linodes" />
+                        <Route
+                          component={PlacementGroups}
+                          path="/placement-groups"
+                        />
                         <Route component={Volumes} path="/volumes" />
                         <Redirect path="/volumes*" to="/volumes" />
-                        {flags.aglb && (
+                        {isACLBEnabled && (
                           <Route
                             component={LoadBalancers}
                             path="/loadbalancer*"