+++ packages/manager/src/features/ObjectStorage/AccessKeyLanding/HostNamesDrawer.test.tsx
@@ -1,0 +1,80 @@
+import { screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import React from 'react';
+
+import { regionFactory } from 'src/factories/regions';
+import { renderWithTheme } from 'src/utilities/testHelpers';
+
+import { HostNamesDrawer } from './HostNamesDrawer';
+
+// Mock the onClose function
+const mockOnClose = vi.fn();
+
+// Mock regions data
+const mockS3Regions = [
+  {
+    id: 'region1',
+    s3_endpoint: 'endpoint1',
+  },
+  {
+    id: 'region2',
+    s3_endpoint: 'endpoint2',
+  },
+];
+
+vi.mock('src/queries/regions', () => ({
+  useRegionsQuery: vi.fn(() => ({
+    data: [
+      ...regionFactory.buildList(1, { id: 'region1', label: 'Newark, NJ' }),
+      ...regionFactory.buildList(1, { id: 'region2', label: 'Atlanta, GA' }),
+    ],
+  })),
+}));
+
+describe('HostNamesDrawer', () => {
+  it('renders the drawer with regions and copyable text', () => {
+    renderWithTheme(
+      <HostNamesDrawer
+        onClose={mockOnClose}
+        open={true}
+        regions={mockS3Regions}
+      />
+    );
+
+    expect(
+      screen.getByRole('dialog', { name: 'Regions / S3 Hostnames' })
+    ).toBeInTheDocument();
+
+    expect(
+      screen.getByRole('heading', { name: 'Regions / S3 Hostnames' })
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('textbox', { name: 'region1: endpoint1' })
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('button', {
+        name: 'Copy S3 Endpoint: Atlanta, GA: endpoint2 to clipboard',
+      })
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('button', {
+        name: 'Copy S3 Endpoint: Newark, NJ: endpoint1 to clipboard',
+      })
+    ).toBeInTheDocument();
+  });
+
+  it('calls onClose when the drawer is closed', () => {
+    renderWithTheme(
+      <HostNamesDrawer
+        onClose={mockOnClose}
+        open={true}
+        regions={mockS3Regions}
+      />
+    );
+
+    const closeButton = screen.getByRole('button', { name: 'Close drawer' });
+    userEvent.click(closeButton);
+
+    expect(mockOnClose).toHaveBeenCalled();
+  });
+});